added cobra based CLI foundation
