+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/wim-srl/container-lab
+module github.com/srl-wim/container-lab
 
 go 1.15
 
@@ -6,7 +6,7 @@ require (
 	docker.io/go-docker v1.0.0
 	github.com/Masterminds/goutils v1.1.0 // indirect
 	github.com/Masterminds/semver v1.5.0 // indirect
-	github.com/Masterminds/sprig v2.22.0+incompatible
+	github.com/Masterminds/sprig v2.22.0+incompatible // indirect
 	github.com/Microsoft/go-winio v0.4.14 // indirect
 	github.com/awalterschulze/gographviz v2.0.1+incompatible
 	github.com/docker/distribution v2.7.1+incompatible // indirect
@@ -19,9 +19,9 @@ require (
 	github.com/mitchellh/copystructure v1.0.0 // indirect
 	github.com/opencontainers/go-digest v1.0.0 // indirect
 	github.com/opencontainers/image-spec v1.0.1 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/sirupsen/logrus v1.6.0
-	github.com/spf13/pflag v1.0.5
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
 	gopkg.in/yaml.v2 v2.3.0
 )
